Rename NonDet.NonEmpty to Choose.
