+++ fused-effects.cabal
@@ -30,6 +30,7 @@ library
   import:              common
   exposed-modules:     Control.Effect
                      , Control.Effect.Carrier
+                     , Control.Effect.Choose
                      , Control.Effect.Cull
                      , Control.Effect.Cut
                      , Control.Effect.Error
@@ -39,7 +40,6 @@ library
                      , Control.Effect.Interpret
                      , Control.Effect.Lift
                      , Control.Effect.NonDet
-                     , Control.Effect.NonDet.NonEmpty
                      , Control.Effect.Pure
                      , Control.Effect.Random
                      , Control.Effect.Reader
